#1535 Jars with given size
* Changing int to long for file size

+++ strongbox-storage/strongbox-storage-api/src/test/java/org/carlspring/strongbox/testing/artifact/TestArtifact.java
@@ -71,6 +71,6 @@ public @interface TestArtifact
     /**
      * Artifact size in bytes.
      */
-    int size() default 1024;
+    long size() default 1024;
 
 }
