+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-maven-layout/strongbox-storage-maven-layout-provider/src/test/java/org/carlspring/strongbox/testing/artifact/MavenTestArtifact.java
@@ -70,5 +70,5 @@ public @interface MavenTestArtifact
      * Additional artifact size in bytes.
      */
     @AliasFor(annotation = TestArtifact.class)
-    int size() default 1024;
+    long size() default 1024;
 }
